Move tests into src/tests/reorder
