+++ .tabignore
@@ -1,6 +1,6 @@
 Changelog
 docs/Makefile
 docs/make.bat
-tests/fixtures/samples/panasonic.jpg
-tests/fixtures/samples/pixel2.jpg
-tests/fixtures/samples/pixel5a.jpg
+src/tests/reorder/fixtures/samples/panasonic.jpg
+src/tests/reorder/fixtures/samples/pixel2.jpg
+src/tests/reorder/fixtures/samples/pixel5a.jpg
