+++ pyproject.toml
@@ -10,7 +10,7 @@ include = [
    { path='NOTICE', format='sdist' },
    { path='LICENSE', format='sdist' },
    { path='README.md', format='sdist' },
-   { path='tests', format='sdist' },
+   { path='src/tests', format='sdist' },
 ]
 classifiers = [
    "Operating System :: Unix",
@@ -109,7 +109,7 @@ pretty = true
 show_absolute_path = true
 show_column_numbers = true
 show_error_codes = true
-files = [ "src/reorder", "tests" ]
+files = [ "src" ]
 check_untyped_defs = true
 disallow_any_generics = true
 disallow_incomplete_defs = true
